Add dedicated packages for component controllers
