+++ controllers/components/dashboard/dashboard_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package components
+package dashboard
 
 import (
 	"context"
