+++ controllers/components/datasciencepipelines/datasciencepipelines_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package components
+package datasciencepipelines
 
 import (
 	"context"
