+++ controllers/components/kserve/kserve_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package components
+package kserve
 
 import (
 	"context"
@@ -27,26 +27,26 @@ import (
 	componentsv1 "github.com/opendatahub-io/opendatahub-operator/v2/apis/components/v1"
 )
 
-// CodeFlareReconciler reconciles a CodeFlare object.
-type CodeFlareReconciler struct {
+// KserveReconciler reconciles a Kserve object.
+type KserveReconciler struct {
 	client.Client
 	Scheme *runtime.Scheme
 }
 
-//+kubebuilder:rbac:groups=components.opendatahub.io,resources=codeflares,verbs=get;list;watch;create;update;patch;delete
-//+kubebuilder:rbac:groups=components.opendatahub.io,resources=codeflares/status,verbs=get;update;patch
-//+kubebuilder:rbac:groups=components.opendatahub.io,resources=codeflares/finalizers,verbs=update
+//+kubebuilder:rbac:groups=components.opendatahub.io,resources=kserves,verbs=get;list;watch;create;update;patch;delete
+//+kubebuilder:rbac:groups=components.opendatahub.io,resources=kserves/status,verbs=get;update;patch
+//+kubebuilder:rbac:groups=components.opendatahub.io,resources=kserves/finalizers,verbs=update
 
 // Reconcile is part of the main kubernetes reconciliation loop which aims to
 // move the current state of the cluster closer to the desired state.
 // TODO(user): Modify the Reconcile function to compare the state specified by
-// the CodeFlare object against the actual cluster state, and then
+// the Kserve object against the actual cluster state, and then
 // perform operations to make the cluster state reflect the state specified by
 // the user.
 //
 // For more details, check Reconcile and its Result here:
 // - https://pkg.go.dev/sigs.k8s.io/controller-runtime@v0.12.2/pkg/reconcile
-func (r *CodeFlareReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+func (r *KserveReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	_ = log.FromContext(ctx)
 
 	// TODO(user): your logic here
@@ -55,8 +55,8 @@ func (r *CodeFlareReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 }
 
 // SetupWithManager sets up the controller with the Manager.
-func (r *CodeFlareReconciler) SetupWithManager(mgr ctrl.Manager) error {
+func (r *KserveReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
-		For(&componentsv1.CodeFlare{}).
+		For(&componentsv1.Kserve{}).
 		Complete(r)
 }
